Move reservation-form to shared folder
Refs #145.
