+++ app/pages/resource/resource-page.less
@@ -1,5 +1,4 @@
 @import 'info/resource-info';
-@import 'reservation-form/reservation-form';
 
 .resource-page {
   .help-text {
