+++ app/shared/reservation-form/catering/CateringOrderTable.spec.js
@@ -7,7 +7,7 @@ import simple from 'simple-mock';
 
 import CateringOrderTable from './CateringOrderTable';
 
-describe('pages/resource/reservation-form/catering/CateringOrderTable', () => {
+describe('shared/reservation-form/catering/CateringOrderTable', () => {
   function getWrapper(props) {
     const defaults = {
       items: [{ id: 'cmi-1', price: 1, quantity: 1 }],
