+++ app/shared/reservation-form/catering/CateringSectionContainer.js
