+++ app/shared/reservation-form/catering/CateringSectionContainer.spec.js
@@ -10,7 +10,7 @@ import {
   UnconnectedCateringSectionContainer as CateringSectionContainer,
 } from './CateringSectionContainer';
 
-describe('pages/resource/reservation-form/catering/CateringSectionContainer', () => {
+describe('shared/reservation-form/catering/CateringSectionContainer', () => {
   function getWrapper(props) {
     const defaults = {
       cateringTime: '10:00',
