+++ app/shared/reservation-form/catering/catering-form/CateringMenuItem.js
