+++ app/shared/reservation-form/catering/catering-modal.less
