+++ app/shared/shared.less
@@ -7,4 +7,5 @@
 @import './modals/reservation-cancel/reservation-cancel';
 @import './modals/reservation-info/reservation-info';
 @import './navbar/navbar';
+@import './reservation-form/reservation-form';
 @import './wrapped-text/wrapped-text';
